Moved most of the methods to useGameLogic. Need to add better comments.
